relocated style files
