+++ mobile-rider/app/(auth)/loginStyles.ts
