+++ mobile-rider/app/(auth)/registerStyles.ts
