+++ mobile-rider/app/(auth)/splashStyles.ts
