+++ mobile-rider/app/(tabs)/IndexStyles.ts
