+++ mobile-rider/app/(tabs)/finesStyles.ts
