+++ mobile-rider/app/(tabs)/qrCodeStyles.ts
