+++ mobile-rider/app/other/settingsStyles.ts
