+++ mobile-rider/app/payments/payment.tsx
@@ -14,6 +14,7 @@ import Icon from "react-native-vector-icons/FontAwesome";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { X } from "lucide-react-native"
 import { useRouter } from "expo-router";
+import styles from "./paymentStyles";
 
 const PaymentMethodCard: React.FC<{
   title: string;
@@ -294,243 +295,5 @@ const PaymentsScreen: React.FC = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    backgroundColor: COLORS.primary,
-  },
-  container: {
-    flex: 1,
-    backgroundColor: COLORS.background,
-  },
-  contentContainer: {
-    padding: 16,
-  },
-  header: {
-    marginBottom: 24,
-  },
-  screenTitle: {
-    fontSize: 28,
-    fontWeight: "bold",
-    color: COLORS.text,
-  },
-  fineSummary: {
-    marginTop: 16,
-  },
-  fineAmount: {
-    fontSize: 20,
-    fontWeight: "bold",
-    color: COLORS.text,
-  },
-  fineDescription: {
-    fontSize: 14,
-    color: "#6B7280",
-    marginTop: 4,
-  },
-  paymentMethods: {
-    marginBottom: 24,
-  },
-  sectionTitle: {
-    fontSize: 18,
-    fontWeight: "600",
-    color: COLORS.text,
-    marginBottom: 16,
-  },
-  paymentMethodsRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  paymentMethodCard: {
-    backgroundColor: COLORS.white,
-    borderRadius: 5,
-    padding: 16,
-    width: "31%",
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.05,
-    shadowRadius: 3,
-    elevation: 2,
-  },
-  paymentMethodCardActive: {
-    backgroundColor: COLORS.primary,
-  },
-  paymentMethodIconContainer: {
-    backgroundColor: COLORS.background,
-    borderRadius: 8,
-    width: 48,
-    height: 48,
-    justifyContent: "center",
-    alignItems: "center",
-    marginBottom: 8,
-  },
-  paymentMethodTitle: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: COLORS.text,
-  },
-  paymentForm: {
-    backgroundColor: COLORS.white,
-    borderRadius: 5,
-    padding: 16,
-    marginBottom: 24,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.05,
-    shadowRadius: 3,
-    elevation: 2,
-  },
-  paymentFormLabel: {
-    fontSize: 16,
-    fontWeight: "600",
-    color: COLORS.text,
-    marginBottom: 16,
-  },
-  inputGroup: {
-    marginBottom: 16,
-  },
-  inputLabel: {
-    fontSize: 14,
-    color: "#6B7280",
-    marginBottom: 8,
-  },
-  input: {
-    backgroundColor: COLORS.background,
-    borderRadius: 8,
-    borderWidth: 1,
-    borderColor: "#E5E7EB",
-    padding: 12,
-    fontSize: 16,
-    color: COLORS.text,
-  },
-  horizontalInputs: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  payButton: {
-    borderRadius: 8,
-    padding: 16,
-    alignItems: "center",
-    marginTop: 8,
-  },
-  payButtonText: {
-    color: COLORS.white,
-    fontSize: 16,
-    fontWeight: "600",
-  },
-  bankDetails: {
-    backgroundColor: COLORS.background,
-    borderRadius: 8,
-    padding: 16,
-    marginBottom: 16,
-  },
-  bankDetailsTitle: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: COLORS.text,
-    marginBottom: 12,
-  },
-  bankDetailRow: {
-    flexDirection: "row",
-    marginBottom: 8,
-  },
-  bankDetailLabel: {
-    fontSize: 14,
-    color: "#6B7280",
-    width: 110,
-  },
-  bankDetailValue: {
-    fontSize: 14,
-    color: COLORS.text,
-    fontWeight: "500",
-    flex: 1,
-  },
-  uploadSection: {
-    marginTop: 16,
-  },
-  uploadTitle: {
-    fontSize: 16,
-    fontWeight: "500",
-    color: COLORS.text,
-    marginBottom: 12,
-  },
-  uploadButton: {
-    borderWidth: 1,
-    borderColor: "#E5E7EB",
-    borderStyle: "dashed",
-    borderRadius: 8,
-    padding: 16,
-    flexDirection: "row",
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: COLORS.background,
-  },
-  uploadButtonText: {
-    fontSize: 14,
-    color: COLORS.text,
-    marginLeft: 8,
-  },
-  receiptPreview: {
-    marginTop: 12,
-    flexDirection: "row",
-    alignItems: "center",
-  },
-  receiptImage: {
-    width: 60,
-    height: 60,
-    borderRadius: 4,
-  },
-  receiptInfo: {
-    marginLeft: 12,
-    flex: 1,
-  },
-  receiptName: {
-    fontSize: 14,
-    color: COLORS.text,
-    marginBottom: 4,
-  },
-  removeReceipt: {
-    fontSize: 12,
-    color: COLORS.accent,
-  },
-  walletSelection: {
-    backgroundColor: COLORS.background,
-    borderRadius: 8,
-    padding: 16,
-  },
-  walletTitle: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: COLORS.text,
-    marginBottom: 12,
-  },
-  walletOption: {
-    flexDirection: "row",
-    alignItems: "center",
-    backgroundColor: COLORS.white,
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 8,
-    borderWidth: 1,
-    borderColor: "#E5E7EB",
-  },
-  walletLogo: {
-    width: 32,
-    height: 32,
-    marginRight: 12,
-  },
-  walletName: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: COLORS.text,
-    flex: 1,
-  },
-  walletNote: {
-    fontSize: 12,
-    color: "#6B7280",
-    marginTop: 12,
-    textAlign: "center",
-  },
-});
 
 export default PaymentsScreen;
