+++ mobile-rider/app/payments/paymentStyles.ts
@@ -16,6 +16,10 @@ export const COLORS = {
 };
 
 const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    backgroundColor: COLORS.primary,
+  },
   container: {
     flex: 1,
     backgroundColor: COLORS.background,
@@ -28,21 +32,20 @@ const styles = StyleSheet.create({
   },
   screenTitle: {
     fontSize: 28,
-    fontWeight: 'bold',
+    fontWeight: "bold",
     color: COLORS.text,
-    marginBottom: 8,
   },
   fineSummary: {
-    marginTop: 8,
+    marginTop: 16,
   },
   fineAmount: {
-    fontSize: 24,
-    fontWeight: 'bold',
+    fontSize: 20,
+    fontWeight: "bold",
     color: COLORS.text,
   },
   fineDescription: {
     fontSize: 14,
-    color: '#6B7280',
+    color: "#6B7280",
     marginTop: 4,
   },
   paymentMethods: {
@@ -50,21 +53,21 @@ const styles = StyleSheet.create({
   },
   sectionTitle: {
     fontSize: 18,
-    fontWeight: '600',
+    fontWeight: "600",
     color: COLORS.text,
     marginBottom: 16,
   },
   paymentMethodsRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
+    flexDirection: "row",
+    justifyContent: "space-between",
   },
   paymentMethodCard: {
     backgroundColor: COLORS.white,
-    borderRadius: 12,
+    borderRadius: 5,
     padding: 16,
-    width: '31%',
-    alignItems: 'center',
-    shadowColor: '#000',
+    width: "31%",
+    alignItems: "center",
+    shadowColor: "#000",
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.05,
     shadowRadius: 3,
@@ -78,21 +81,21 @@ const styles = StyleSheet.create({
     borderRadius: 8,
     width: 48,
     height: 48,
-    justifyContent: 'center',
-    alignItems: 'center',
+    justifyContent: "center",
+    alignItems: "center",
     marginBottom: 8,
   },
   paymentMethodTitle: {
     fontSize: 14,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.text,
   },
   paymentForm: {
     backgroundColor: COLORS.white,
-    borderRadius: 12,
+    borderRadius: 5,
     padding: 16,
     marginBottom: 24,
-    shadowColor: '#000',
+    shadowColor: "#000",
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.05,
     shadowRadius: 3,
@@ -100,7 +103,7 @@ const styles = StyleSheet.create({
   },
   paymentFormLabel: {
     fontSize: 16,
-    fontWeight: '600',
+    fontWeight: "600",
     color: COLORS.text,
     marginBottom: 16,
   },
@@ -109,32 +112,32 @@ const styles = StyleSheet.create({
   },
   inputLabel: {
     fontSize: 14,
-    color: '#6B7280',
+    color: "#6B7280",
     marginBottom: 8,
   },
   input: {
     backgroundColor: COLORS.background,
     borderRadius: 8,
     borderWidth: 1,
-    borderColor: '#E5E7EB',
+    borderColor: "#E5E7EB",
     padding: 12,
     fontSize: 16,
     color: COLORS.text,
   },
   horizontalInputs: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
+    flexDirection: "row",
+    justifyContent: "space-between",
   },
   payButton: {
     borderRadius: 8,
     padding: 16,
-    alignItems: 'center',
+    alignItems: "center",
     marginTop: 8,
   },
   payButtonText: {
     color: COLORS.white,
     fontSize: 16,
-    fontWeight: '600',
+    fontWeight: "600",
   },
   bankDetails: {
     backgroundColor: COLORS.background,
@@ -144,23 +147,23 @@ const styles = StyleSheet.create({
   },
   bankDetailsTitle: {
     fontSize: 14,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.text,
     marginBottom: 12,
   },
   bankDetailRow: {
-    flexDirection: 'row',
+    flexDirection: "row",
     marginBottom: 8,
   },
   bankDetailLabel: {
     fontSize: 14,
-    color: '#6B7280',
+    color: "#6B7280",
     width: 110,
   },
   bankDetailValue: {
     fontSize: 14,
     color: COLORS.text,
-    fontWeight: '500',
+    fontWeight: "500",
     flex: 1,
   },
   uploadSection: {
@@ -168,19 +171,19 @@ const styles = StyleSheet.create({
   },
   uploadTitle: {
     fontSize: 16,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.text,
     marginBottom: 12,
   },
   uploadButton: {
     borderWidth: 1,
-    borderColor: '#E5E7EB',
-    borderStyle: 'dashed',
+    borderColor: "#E5E7EB",
+    borderStyle: "dashed",
     borderRadius: 8,
     padding: 16,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
     backgroundColor: COLORS.background,
   },
   uploadButtonText: {
@@ -190,8 +193,8 @@ const styles = StyleSheet.create({
   },
   receiptPreview: {
     marginTop: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
+    flexDirection: "row",
+    alignItems: "center",
   },
   receiptImage: {
     width: 60,
@@ -211,73 +214,6 @@ const styles = StyleSheet.create({
     fontSize: 12,
     color: COLORS.accent,
   },
-  transactionHistory: {
-    marginBottom: 16,
-  },
-  transactionHistoryHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 12,
-  },
-  viewAllButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  viewAllText: {
-    fontSize: 14,
-    color: COLORS.primary,
-    marginRight: 4,
-  },
-  transactionItem: {
-    backgroundColor: COLORS.white,
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  transactionHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  transactionInfo: {
-    flex: 1,
-  },
-  transactionDescription: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: COLORS.text,
-    marginBottom: 4,
-  },
-  transactionMeta: {
-    fontSize: 12,
-    color: '#6B7280',
-  },
-  transactionAmount: {
-    alignItems: 'flex-end',
-  },
-  transactionAmountText: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: COLORS.text,
-    marginBottom: 4,
-  },
-  transactionStatus: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-    borderRadius: 12,
-  },
-  transactionStatusText: {
-    fontSize: 10,
-    fontWeight: '500',
-    marginLeft: 4,
-  },
   walletSelection: {
     backgroundColor: COLORS.background,
     borderRadius: 8,
@@ -285,19 +221,19 @@ const styles = StyleSheet.create({
   },
   walletTitle: {
     fontSize: 14,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.text,
     marginBottom: 12,
   },
   walletOption: {
-    flexDirection: 'row',
-    alignItems: 'center',
+    flexDirection: "row",
+    alignItems: "center",
     backgroundColor: COLORS.white,
     borderRadius: 8,
     padding: 12,
     marginBottom: 8,
     borderWidth: 1,
-    borderColor: '#E5E7EB',
+    borderColor: "#E5E7EB",
   },
   walletLogo: {
     width: 32,
@@ -306,15 +242,15 @@ const styles = StyleSheet.create({
   },
   walletName: {
     fontSize: 14,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.text,
     flex: 1,
   },
   walletNote: {
     fontSize: 12,
-    color: '#6B7280',
+    color: "#6B7280",
     marginTop: 12,
-    textAlign: 'center',
+    textAlign: "center",
   },
 });
